// src/components/custom/stepper/type.ts
export type FormData = {
  firstName: string;
  lastName: string;
  company: string;
  role: string;
  type: any;
  markets: string;
  email: string;
  password: any;
  invitedBy: string;
  affiliateFriends: any | null;
  socialPlatforms: any[];
  affiliateLink: string;
  birthDate: string;
  location?: any;
  gender?: any;
  diseaseAreas?: any[];
  experienceAs?: any;
  ethnicity?: any;
  instaP: string;
  instaR: string;
  instaS: string;
  yVideoS: any;
  yVideoM: any;
  yVideoL: any;
  ttPost: any;
  currency: number;
  questionCredit: string;
  averageQuestionSurvey: string;
  interviewShort: string;
  interviewLong: string;
  status: number;
  instagramUsername: string;
  specialization?: any;
};